mac/apphub: move mac options into AppHub
