+++ osdep/mac/application.h
@@ -17,33 +17,5 @@
 
 #pragma once
 
-#include "options/m_option.h"
-
-enum {
-    FRAME_VISIBLE = 0,
-    FRAME_WHOLE,
-};
-
-enum {
-    RENDER_TIMER_CALLBACK = 0,
-    RENDER_TIMER_PRECISE,
-    RENDER_TIMER_SYSTEM,
-};
-
-struct macos_opts {
-    int macos_title_bar_appearance;
-    int macos_title_bar_material;
-    struct m_color macos_title_bar_color;
-    int macos_fs_animation_duration;
-    bool macos_force_dedicated_gpu;
-    int macos_app_activation_policy;
-    int macos_geometry_calculation;
-    int macos_render_timer;
-    int cocoa_cb_sw_renderer;
-    bool cocoa_cb_10bit_context;
-};
-
 // multithreaded wrapper for mpv_main
 int cocoa_main(int argc, char *argv[]);
-
-extern const struct m_sub_options macos_conf;
