+++ osdep/mac/application_objc.h
@@ -26,8 +26,6 @@ struct mpv_handle;
 @interface Application : NSApplication
 
 - (void)initCocoaCb:(struct mpv_handle *)ctx;
-+ (const struct m_sub_options *)getMacConf;
-+ (const struct m_sub_options *)getVoConf;
 
 @property(nonatomic, retain) MenuBar *menuBar;
 @property(nonatomic, assign) size_t openCount;
